+++ vnext/Microsoft.ReactNative/Views/ICompositionRootView.h
@@ -12,7 +12,6 @@
 namespace Microsoft::ReactNative {
 
 struct ICompositionRootView : public facebook::react::IReactRootView {
-  virtual winrt::Microsoft::ReactNative::Composition::IVisual GetVisual() const noexcept = 0;
   virtual float ScaleFactor() noexcept = 0;
 };